requirement.txt
Flask
mysql_connector_python
requests
jsonify
